backend/requirements.txt
fastapi[standard]
pydantic
python-dotenv
